feat: add divider component (#10)
* feat: add divider component

+++ scripts/rename-db-ui-base.js
@@ -1,6 +1,14 @@
 const FS = require('fs-extra');
 
-FS.moveSync(
-	'../packages/foundations/scss/db-ui-base.scss',
-	'../packages/foundations/scss/db-ui-foundations.scss'
-);
+if (FS.pathExistsSync('./packages/foundations/scss/db-ui-base.scss')) {
+	if (
+		FS.pathExistsSync('./packages/foundations/scss/db-ui-foundations.scss')
+	) {
+		FS.removeSync('./packages/foundations/scss/db-ui-foundations.scss');
+	}
+
+	FS.moveSync(
+		'./packages/foundations/scss/db-ui-base.scss',
+		'./packages/foundations/scss/db-ui-foundations.scss'
+	);
+}
